@@ -11,7 +11,7 @@ from sqlalchemy.dialects import postgresql
 
 # revision identifiers, used by Alembic.
 revision = 'aeb085e011f4'
-down_revision = '2024_08_31_character_safeguards'
+down_revision = '2024_08_31_character_safeguards'  # Depends on character safeguards
 branch_labels = None
 depends_on = None
 
