@@ -1,16 +1,21 @@
 """Character API Tests"""
 
-from fastapi.testclient import TestClient
+import pytest
+from uuid import uuid4
+from httpx import AsyncClient
 
-def create_test_character(client: TestClient, name: str = "Test Character") -> dict:
+async def create_test_character(client: AsyncClient, name: str = "Test Character") -> dict:
     """Helper to create a test character via HTTP."""
     character_data = {
         "name": name,
-        "race": "Human",
-        "class_": "Fighter",
-        "background": "Folk Hero",
-        "alignment": "Lawful Good",
-        "data": {
+        "user_id": str(uuid4()),
+        "campaign_id": str(uuid4()),
+        "character_data": {
+            "species": "Human",
+            "background": "Folk Hero",
+            "level": 1,
+            "class_": "Fighter",
+            "alignment": "Lawful Good",
             "abilities": {
                 "strength": 15,
                 "dexterity": 14,
@@ -23,78 +28,83 @@ def create_test_character(client: TestClient, name: str = "Test Character") -> d
             "equipment": {"weapons": ["Longsword", "Shield"]}
         }
     }
-    response = client.post("/api/v2/characters", json=character_data)
+    response = await client.post("/api/v2/characters", json=character_data)
     assert response.status_code == 200
     return response.json()
 
-def test_list_characters(client: TestClient):
+@pytest.mark.asyncio
+async def test_list_characters(client: AsyncClient):
     """Test listing characters."""
     # Create a test character first
-    character = create_test_character(client)
+    character = await create_test_character(client)
     
     # Test listing
-    response = client.get("/api/v2/characters")
+    response = await client.get("/api/v2/characters")
     assert response.status_code == 200
     data = response.json()
     assert isinstance(data, list)
     assert len(data) > 0
     assert any(c["id"] == character["id"] for c in data)
 
-def test_get_character(client: TestClient):
+@pytest.mark.asyncio
+async def test_get_character(client: AsyncClient):
     """Test getting a specific character."""
     # Create a test character first
-    character = create_test_character(client)
+    character = await create_test_character(client)
     
     # Test retrieval
-    response = client.get(f"/api/v2/characters/{character['id']}")
+    response = await client.get(f"/api/v2/characters/{character['id']}")
     assert response.status_code == 200
     data = response.json()
     assert data["name"] == character["name"]
-    assert data["user_id"] == character["user_id"]
+    assert isinstance(data["user_id"], str)  # Should be a UUID string
 
-def test_create_character(client: TestClient):
+@pytest.mark.asyncio
+async def test_create_character(client: AsyncClient):
     """Test character creation."""
-    character = create_test_character(client, name="New Character")
+    character = await create_test_character(client, name="New Character")
     assert character["name"] == "New Character"
-    assert character["user_id"] == "test_user"
+    assert isinstance(character["user_id"], str)  # Should be a UUID string
+    assert isinstance(character["campaign_id"], str)  # Should be a UUID string
     assert character["character_data"]["species"] == "Human"
 
-def test_update_character(client: TestClient):
+@pytest.mark.asyncio
+async def test_update_character(client: AsyncClient):
     """Test character update."""
     # Create a test character first
-    character = create_test_character(client)
+    character = await create_test_character(client)
     
     # Test update
     update_data = {
         "name": "Updated Character",
-        "race": character["race"],
-        "class_": character["class_"],
-        "background": character["background"],
-        "alignment": character["alignment"],
-        "data": character["data"]
+        "user_id": character["user_id"],
+        "campaign_id": character["campaign_id"],
+        "character_data": character["character_data"]
     }
-    response = client.put(f"/api/v2/characters/{character['id']}", json=update_data)
+    response = await client.put(f"/api/v2/characters/{character['id']}", json=update_data)
     assert response.status_code == 200
     data = response.json()
     assert data["name"] == "Updated Character"
 
-def test_delete_character(client: TestClient):
+@pytest.mark.asyncio
+async def test_delete_character(client: AsyncClient):
     """Test character deletion."""
     # Create a test character first
-    character = create_test_character(client)
+    character = await create_test_character(client)
     
     # Test deletion
-    response = client.delete(f"/api/v2/characters/{character['id']}")
+    response = await client.delete(f"/api/v2/characters/{character['id']}")
     assert response.status_code == 200
     
     # Verify character is deleted
-    response = client.get(f"/api/v2/characters/{character['id']}")
+    response = await client.get(f"/api/v2/characters/{character['id']}")
     assert response.status_code == 404
 
-def test_direct_edit_character(client: TestClient):
+@pytest.mark.asyncio
+async def test_direct_edit_character(client: AsyncClient):
     """Test direct character edit."""
     # Create a test character first
-    character = create_test_character(client)
+    character = await create_test_character(client)
     
     # Test direct edit
     edit_data = {
@@ -108,9 +118,9 @@ def test_direct_edit_character(client: TestClient):
         },
         "notes": "Testing direct edit functionality"
     }
-    response = client.post(f"/api/v2/characters/{character['id']}/direct-edit", json=edit_data)
+    response = await client.post(f"/api/v2/characters/{character['id']}/direct-edit", json=edit_data)
     assert response.status_code == 200
     data = response.json()
     assert data["name"] == "Directly Edited Character"
-    assert data["level"] == 2  # Level is now a top-level field
-    assert data["user_modified"] is True
+    assert data["character_data"]["level"] == 2  # Level is in character_data
+    assert data["character_data"]["character_classes"]["Fighter"] == 2
