@@ -1,10 +1,13 @@
 """Repository Tests"""
 
+import uuid
 import pytest
+from datetime import datetime
 from sqlalchemy.ext.asyncio import AsyncSession
 from character_service.repositories.character_repository import CharacterRepository
 from character_service.repositories.journal_repository import JournalRepository
 from character_service.repositories.inventory_repository import InventoryRepository
+from character_service.models.models import Character
 from character_service.schemas.schemas import (
     CharacterCreate,
     CharacterUpdate,
@@ -14,12 +17,8 @@ from character_service.schemas.schemas import (
     InventoryItemUpdate
 )
 
-@pytest.mark.asyncio
-async def test_character_repository(test_db: AsyncSession):
-    """Test character repository operations."""
-    repo = CharacterRepository(test_db)
-
-    # Test create
+async def create_test_character(repo: CharacterRepository) -> Character:
+    """Helper to create a test character via repository."""
     character_data = {
         "name": "Test Character",
         "user_id": "test_user",
@@ -27,10 +26,29 @@ async def test_character_repository(test_db: AsyncSession):
         "character_data": {
             "species": "Human",
             "background": "Folk Hero",
-            "level": 1
+            "level": 1,
+            "class_": "Fighter",
+            "alignment": "Lawful Good",
+            "abilities": {
+                "strength": 15,
+                "dexterity": 14,
+                "constitution": 13,
+                "intelligence": 12,
+                "wisdom": 10,
+                "charisma": 8
+            },
+            "skills": ["Athletics", "Intimidation"],
+            "equipment": {"weapons": ["Longsword", "Shield"]}
         }
     }
-    character = await repo.create(CharacterCreate(**character_data))
+    return await repo.create(CharacterCreate(**character_data))
+@pytest.mark.asyncio
+async def test_character_repository(db_session: AsyncSession):
+    """Test character repository operations."""
+    repo = CharacterRepository(db_session)
+
+    # Test create
+    character = await create_test_character(repo)
     assert character.name == "Test Character"
     assert character.user_id == "test_user"
 
@@ -54,20 +72,24 @@ async def test_character_repository(test_db: AsyncSession):
     assert await repo.get(character.id) is None
 
 @pytest.mark.asyncio
-async def test_journal_repository(test_db: AsyncSession, test_character):
+async def test_journal_repository(db_session: AsyncSession):
     """Test journal repository operations."""
-    repo = JournalRepository(test_db)
+    char_repo = CharacterRepository(db_session)
+    character = await create_test_character(char_repo)
+
+    repo = JournalRepository(db_session)
 
     # Test create
     entry_data = JournalEntryCreate(
-        character_id=test_character.id,
+        character_id=character.id,
+        timestamp=datetime.utcnow(),
         title="Test Entry",
         content="Test content",
         entry_type="session"
     )
     entry = await repo.create(entry_data)
     assert entry.title == "Test Entry"
-    assert entry.character_id == test_character.id
+    assert entry.character_id == character.id
 
     # Test get
     retrieved = await repo.get(entry.id)
@@ -75,9 +97,9 @@ async def test_journal_repository(test_db: AsyncSession, test_character):
     assert retrieved.title == entry.title
 
     # Test get by character
-    entries = await repo.get_all_by_character(test_character.id)
+    entries = await repo.get_all_by_character(character.id)
     assert len(entries) > 0
-    assert entries[0].character_id == test_character.id
+    assert entries[0].character_id == character.id
 
     # Test update
     update_data = JournalEntryUpdate(title="Updated Entry")
@@ -89,13 +111,16 @@ async def test_journal_repository(test_db: AsyncSession, test_character):
     assert await repo.get(entry.id) is None
 
 @pytest.mark.asyncio
-async def test_inventory_repository(test_db: AsyncSession, test_character):
+async def test_inventory_repository(db_session: AsyncSession):
     """Test inventory repository operations."""
-    repo = InventoryRepository(test_db)
+    char_repo = CharacterRepository(db_session)
+    character = await create_test_character(char_repo)
+
+    repo = InventoryRepository(db_session)
 
     # Test create
     item_data = InventoryItemCreate(
-        character_id=test_character.id,
+        character_id=character.id,
         item_data={
             "name": "Test Sword",
             "type": "weapon",
@@ -105,7 +130,7 @@ async def test_inventory_repository(test_db: AsyncSession, test_character):
     )
     item = await repo.create(item_data)
     assert item.item_data["name"] == "Test Sword"
-    assert item.character_id == test_character.id
+    assert item.character_id == character.id
 
     # Test get
     retrieved = await repo.get(item.id)
@@ -113,9 +138,9 @@ async def test_inventory_repository(test_db: AsyncSession, test_character):
     assert retrieved.item_data["name"] == "Test Sword"
 
     # Test get by character
-    items = await repo.get_all_by_character(test_character.id)
+    items = await repo.get_all_by_character(character.id)
     assert len(items) > 0
-    assert items[0].character_id == test_character.id
+    assert items[0].character_id == character.id
 
     # Test update
     update_data = InventoryItemUpdate(
@@ -135,11 +160,11 @@ async def test_inventory_repository(test_db: AsyncSession, test_character):
     assert await repo.get(item.id) is None
 
 @pytest.mark.asyncio
-async def test_repository_error_handling(test_db: AsyncSession):
+async def test_repository_error_handling(db_session: AsyncSession):
     """Test repository error handling."""
-    char_repo = CharacterRepository(test_db)
-    journal_repo = JournalRepository(test_db)
-    inventory_repo = InventoryRepository(test_db)
+    char_repo = CharacterRepository(db_session)
+    journal_repo = JournalRepository(db_session)
+    inventory_repo = InventoryRepository(db_session)
 
     # Test get non-existent items
     assert await char_repo.get(999999) is None
